+++ src/App.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
 import { TranslationProvider } from 'modules/translation/container';
 import RoutesApp from 'routes';
+import RoutesProvider from 'routes/container';
 import './App.css';
 
 const App: React.FC = () => {
   return (
     <TranslationProvider>
-      <RoutesApp />
+      <RoutesProvider>
+        <RoutesApp />
+      </RoutesProvider>
     </TranslationProvider>
   );
 }
